refactor(noscript): move inline scripts to new file
